+++ modules/instances/main.tf
@@ -45,6 +45,8 @@ resource "oci_core_instance" "instances" {
   display_name         = each.value.name
   preserve_boot_volume = lookup(each.value.optionals, "preserve_boot_volume", true)
   state                = each.value.state
+  freeform_tags        = lookup(each.value, "freeform_tags", null)
+  defined_tags         = merge(lookup(each.value, "defined_tags", {}),{ "Oracle-Tags.CreatedBy" = "terraform" })
   metadata = {
     ssh_authorized_keys = each.value.autherized_keys
     user_data           = lookup(each.value.optionals, "user_data", null)
@@ -78,6 +80,7 @@ resource "oci_core_instance" "instances" {
     skip_source_dest_check    = lookup(each.value.optionals, "skip_source_dest_check", false)
     assign_private_dns_record = true
     private_ip                = each.value.config.primary_vnic.primary_ip
+    assign_ipv6ip             = each.value.ipv6
   }
 
   dynamic "source_details" {
@@ -140,6 +143,7 @@ resource "oci_core_vnic_attachment" "secondary_vnic_attachment" {
     subnet_id                 = each.value.vnic.subnet_id
     hostname_label            = each.value.vnic.hostname_label
     skip_source_dest_check    = each.value.vnic.skip_source_dest_check
+    assign_ipv6ip             = lookup(each.value.vnic, "ipv6", false)
   }
 
   instance_id = oci_core_instance.instances[each.value.instance_key].id