+++ modules/kubernetes/main.tf
@@ -24,6 +24,14 @@ resource "oci_containerengine_cluster" "cluster" {
       is_tiller_enabled               = true
     }
     service_lb_subnet_ids = var.lb_subnet_ids
+    ip_families = var.ip_families
+    dynamic "kubernetes_network_config" {
+      for_each = var.kubernetes_network_config == null ? [] : [var.kubernetes_network_config]
+      content {
+        pods_cidr     = try(kubernetes_network_config.value.pods_cidr, null)
+        services_cidr = try(kubernetes_network_config.value.services_cidr, null)
+      }
+    }
   }
 
   lifecycle {