+++ modules/object-storage/main.tf
@@ -6,10 +6,12 @@ resource "oci_objectstorage_bucket" "bucket" {
   namespace             = data.oci_objectstorage_namespace.namespace.namespace
   compartment_id        = each.value.compartment_id
   name                  = each.value.name
+  defined_tags          = merge(lookup(each.value, "defined_tags", {}),{ "Oracle-Tags.CreatedBy" = "terraform" })
+  freeform_tags         = lookup(each.value, "freeform_tags", null)
   access_type           = each.value.is_public == true ? "ObjectReadWithoutList" : "NoPublicAccess"
   storage_tier          = each.value.storage_tier
-  object_events_enabled = lookup(each.value.optionals, "object_events_enabled", false)
-  versioning            = lookup(each.value.optionals, "versioning_enabled", false) ? "Enabled" : "Disabled"
+  object_events_enabled = try(each.value.optionals.object_events_enabled, false)
+  versioning            = try(each.value.optionals.versioning, "Disabled")
 }
 
 resource "oci_objectstorage_object_lifecycle_policy" "lifecycle_policy" {
@@ -45,4 +47,14 @@ resource "oci_objectstorage_object_lifecycle_policy" "lifecycle_policy" {
       time_unit   = rules.value.time_unit
     }
   }
+}
+resource "oci_objectstorage_replication_policy" "bucket_replication" {
+  for_each = {
+    for k, v in var.buckets : k => v if try(v.optionals.replication_policy.source_region_name, null) == var.region
+  }
+  name                    = each.value.name
+  bucket                  = oci_objectstorage_bucket.bucket[each.key].name
+  namespace               = data.oci_objectstorage_namespace.namespace.namespace
+  destination_bucket_name = coalesce(try(each.value.optionals.replication_policy.destination_bucket_name, null), oci_objectstorage_bucket.bucket[each.key].name)
+  destination_region_name = each.value.optionals.replication_policy.destination_region_name
 }